requirements.txt
django-bootstrap4
python-dateutil
django-import-export
django-qr-code